Implementacija zahteva za ocenu 6 pred krajem

+++ src/main/java/com/sergej/web_chocholate_shop/service/ProductService.java
@@ -1,10 +1,15 @@
 package com.sergej.web_chocholate_shop.service;
 
+import com.sergej.web_chocholate_shop.model.entity.Discount;
 import com.sergej.web_chocholate_shop.model.entity.Factory;
 import com.sergej.web_chocholate_shop.repository.ProductRepository;
+import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import com.sergej.web_chocholate_shop.model.entity.Product;
 
+import java.math.BigDecimal;
+import java.math.RoundingMode;
+import java.time.LocalDateTime;
 import java.util.List;
 
 @Service
@@ -110,6 +115,110 @@ public class ProductService {
 
         findById(id);
 
+        Product product = findById(id);
+
+        if(!product.getPurchaseItems().isEmpty()) {
+
+            throw new RuntimeException("Product cannot be deleted because it exists in purchases.");
+        }
+
         productRepository.deleteById(id);
     }
+
+    public BigDecimal calculateCurrentPrice(
+            Product product) {
+
+        Discount discount = product.getDiscount();
+
+        if(discount == null) {
+
+            return product.getPrice();
+        }
+
+        LocalDateTime now = LocalDateTime.now();
+
+        if(now.isBefore(discount.getStartDateTime())) {
+
+            return product.getPrice();
+        }
+
+        if(now.isAfter(discount.getEndDateTime())
+        ) {
+
+            return product.getPrice();
+        }
+
+        BigDecimal percent = BigDecimal.valueOf(
+                        discount.getDiscountPercent());
+
+        BigDecimal multiplier =
+                BigDecimal.valueOf(100)
+                        .subtract(percent)
+                        .divide(BigDecimal.valueOf(100));
+
+        return product.getPrice()
+                .multiply(multiplier).setScale(2,
+                        RoundingMode.HALF_UP);
+    }
+
+    public boolean hasActiveDiscount(
+            Product product) {
+
+        Discount discount = product.getDiscount();
+
+        if(discount == null) {
+            return false;
+        }
+
+        LocalDateTime now = LocalDateTime.now();
+
+        return !now.isBefore(
+                discount.getStartDateTime())
+                &&
+                !now.isAfter(
+                        discount.getEndDateTime());
+    }
+
+    public List<Product> searchProducts(
+            String code,
+            String name,
+            Long factoryId,
+            Integer minStock,
+            Integer maxStock,
+            BigDecimal minPrice,
+            BigDecimal maxPrice,
+            String sortBy,
+            String direction) {
+
+        Sort sort = Sort.unsorted();
+
+        if (sortBy != null && !sortBy.isBlank()) {
+
+            if ("desc".equalsIgnoreCase(direction)) {
+
+                sort = Sort.by(sortBy).descending();
+
+            } else {
+
+                sort = Sort.by(sortBy).ascending();
+
+            }
+        }
+
+        return productRepository.searchProducts(
+                code,
+                name,
+                factoryId,
+                minStock,
+                maxStock,
+                minPrice,
+                maxPrice,
+                sort);
+    }
+
+
+
+
+
+
 }
